+++ tests/testthat/test-center.R
@@ -58,6 +58,17 @@ test_that("center, all NA or Inf", {
   expect_equal(z, c(NA, -Inf, Inf), ignore_attr = TRUE)
 })
 
+test_that("center works correctly with only one value", {
+  expect_equal(center(100), 0, ignore_attr = TRUE)
+  expect_message(center(100), "will be set to 0")
+  expect_equal(center(100, center = 1), 99, ignore_attr = TRUE)
+  expect_equal(
+    center(100, reference = mtcars$mpg),
+    100 - mean(mtcars$mpg),
+    ignore_attr = TRUE
+  )
+})
+
 
 
 # with grouped data -------------------------------------------