+++ NEWS.md
@@ -2,6 +2,11 @@
 
 * Etienne Bacher is the new maintainer.
 
+BUG FIXES
+
+* `center(x)` now works correctly when `x` is a single value and either
+  `reference` or `center` is specified (#324).
+
 # datawizard 0.6.4
 
 NEW FUNCTIONS